requirements.txt
absl-py==0.8.1
asn1crypto==1.2.0
astor==0.8.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
conda==4.7.12
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.8
grpcio==1.16.1
h5py==2.8.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
llvmlite==0.30.0
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numba==0.46.0
numpy==1.17.4
olefile==0.46
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.10.1
pycosat==0.6.3
pycparser==2.19
pyOpenSSL==19.1.0
pyparsing==2.4.5
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
ruamel-yaml==0.15.46
scipy==1.3.1
six==1.13.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
tensorflow-hub==0.7.0
termcolor==1.1.0
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.40.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
